changed package names/structure, changed constructor in curriculum, and made necessary changes to where it was used
