+++ hydra-curriculum-service/src/main/java/com/revature/hydra/curriculum/beans/DaysDTO.java
@@ -1,4 +1,4 @@
-package com.revature.hydra.curriculum.pojos;
+package com.revature.hydra.curriculum.beans;
 
 import java.util.Arrays;
