+++ hydra-curriculum-service/src/main/java/com/revature/hydra/curriculum/beans/Role.java
@@ -1,4 +1,4 @@
-package com.revature.hydra.curriculum.pojos;
+package com.revature.hydra.curriculum.beans;
 
 public enum Role {
     NONE, ASSOCIATE, TRAINER, ADMIN;
