+++ hydra-curriculum-service/src/main/java/com/revature/hydra/curriculum/beans/SubtopicName.java
@@ -1,4 +1,4 @@
-package com.revature.hydra.curriculum.pojos;
+package com.revature.hydra.curriculum.beans;
 
 /**
  * non-persistent bean for Subtopic Name
