+++ hydra-curriculum-service/src/main/java/com/revature/hydra/curriculum/repositories/CurriculumRepository.java
@@ -1,10 +1,10 @@
-package com.revature.hydra.curriculum.repository;
+package com.revature.hydra.curriculum.repositories;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.revature.hydra.curriculum.bean.Curriculum;
+import com.revature.hydra.curriculum.beans.Curriculum;
 
 public interface CurriculumRepository extends JpaRepository<Curriculum, Integer> {
