+++ hydra-curriculum-service/src/main/java/com/revature/hydra/curriculum/services/CurriculumService.java
@@ -1,4 +1,4 @@
-package com.revature.hydra.curriculum.service;
+package com.revature.hydra.curriculum.services;
 
 import java.util.List;
 
@@ -7,9 +7,9 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.revature.hydra.curriculum.bean.Curriculum;
-import com.revature.hydra.curriculum.repository.CurriculumRepository;
-import com.revature.hydra.curriculum.repository.CurriculumSubtopicRepository;
+import com.revature.hydra.curriculum.beans.Curriculum;
+import com.revature.hydra.curriculum.repositories.CurriculumRepository;
+import com.revature.hydra.curriculum.repositories.CurriculumSubtopicRepository;
 
 /**
  * A Service class for retrieving and modifying curriculum data.
